Revert "Move test app to proper package structure. Cleanup code"
This reverts commit c0fbd3671d698d5da3187a8a5cadc9bffc20ccdf.
